requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
gevent==1.2.2
greenlet==0.4.12
holster==2.0.0
idna==2.6
kazoo==2.4.0
msgpack==0.5.1
peewee==2.10.2
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
websocket-client==0.44.0
